pyproject.toml
[project]
name = "puka"
version = "3.1.1"
description = "Where evything finds its place."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-taggit>=6.1.0",
    "django>=5.1.4",
    "django-crispy-forms>=2.3",
    "django-htmx>=1.21.0",
    "environs[django]>=11.2.1",
    "gunicorn>=23.0.0",
    "psycopg[c,pool]>=3.2.3",
    "whitenoise>=6.8.2",
    "django-filter>=24.3",
    "crispy-tailwind>=1.0.3",
    "django-template-partials>=24.4",
    "django-treebeard>=4.7.1",
    "python-dateutil>=2.9.0.post0",
    "types-python-dateutil>=2.9.0.20241206",
]

[project.scripts]
puka-manage = "puka.manage:main"

[dependency-groups]
dev = [
    { include-group = "test" },
    "django-debug-toolbar>=4.4.6",
    "ipython>=8.30.0",
    "rich>=13.9.4",
    "ruff>=0.8.2",
    "pywatchman>=2.0.0",
]
test = [
    "django-stubs[compatible-mypy]>=5.1.1",
    "mypy>=1.13.0",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "pytest-factoryboy>=2.7.0",
]
pre-commit = [
    "add-trailing-comma>=3.1.0",
    "djade>=1.3.2",
    "django-upgrade>=1.23.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = ["/.*", "/docs", "/tests"]

[tool.pytest.ini_options]
testpaths = ["puka"]
python_files = "*_test.py"
DJANGO_SETTINGS_MODULE = "puka.settings.test"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "puka.settings.test"

[tool.ruff]
line-length = 99
exclude = ["puka/*/migrations"]
force-exclude = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN", "COM812", "D1", "D203", "D212", "E501", "ISC001"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["ARG001", "PLR2004", "S101"]
"puka/manage.py" = ["EM101", "TRY003"]
"puka/settings/*.py" = ["S105"]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
pythonVersion = "3.12"
reportUnusedParameter = false
reportIncompatibleVariableOverride = false
reportAttributeAccessIssue = false

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportAttributeAccessIssue = "none"
